cleanup wallet, lockdrop, home
