+++ components/Lockdrop/hooks/useClaimFees.ts
@@ -1,6 +1,5 @@
 import contracts from '@/config/contracts.json'
 import { VestingMsgComposer } from '@/contracts/codegen/vesting/Vesting.message-composer'
-import { decodeMsgs } from '@/helpers/decodeMsg'
 import useSimulateAndBroadcast from '@/hooks/useSimulateAndBroadcast'
 import useWallet from '@/hooks/useWallet'
 import { queryClient } from '@/pages/_app'
